Понастроил связей, сейчас разбираться буду, хотя бы приложение запустилось

+++ src/main/java/kostuchenkov/rgr/repository/OrderRepository.java
@@ -3,6 +3,6 @@ package kostuchenkov.rgr.repository;
 import kostuchenkov.rgr.domain.Order;
 import org.springframework.data.repository.CrudRepository;
 
-public interface OrderRepository extends CrudRepository<Order, Integer> {
+public interface OrderRepository extends CrudRepository<Order, Long> {
 
 }
