+++ src/main/java/kostuchenkov/rgr/repository/ReviewRepository.java
@@ -3,5 +3,5 @@ package kostuchenkov.rgr.repository;
 import kostuchenkov.rgr.domain.Review;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ReviewRepository extends CrudRepository<Review, Integer> {
+public interface ReviewRepository extends CrudRepository<Review, Long> {
 }
